pyproject.toml
[project]
name = "rerun_high_frequency_transforms"
requires-python = ">=3.11"
version = "0.1.0"
readme = "README.md"
dependencies = [
	"rerun-sdk[datafusion]>=0.29.0-rc.2",
	"rosbags"
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = [
	"linux-64",
	"linux-aarch64",
	"osx-arm64",
	"win-64",
]

[tool.pixi.tasks.static_example]
cmd = "python3 example.py --static_transforms"


[tool.pixi.tasks.dynamic_example]
cmd = "python3 example.py"